Ensure we resolve our own internal deps to the same version
