+++ package.json
@@ -17,8 +17,8 @@
     "lint:fix": "eslint . --ext ts,.tsx --fix",
     "test": "NODE_ENV=test jest",
     "test:unit": "yarn test",
-    "upgrade-synthetix": "yarn up $(./scripts/resolve-synthetix-versions.js latest) --exact",
-    "upgrade-synthetix:dev": "yarn up $(./scripts/resolve-synthetix-versions.js dev) --exact",
+    "upgrade-synthetix": "./scripts/upgrade-synthetix-versions.js latest && yarn dedupe",
+    "upgrade-synthetix:dev": "./scripts/upgrade-synthetix-versions.js dev && yarn dedupe",
     "upgrade-browsers": "yarn up caniuse-lite --recursive",
     "deps": "depcheck --config=.depcheckrc.json",
     "storybook": "start-storybook -p 6006",
@@ -31,15 +31,14 @@
     "@emotion/styled": "^11.10.0",
     "@reach/dialog": "^0.15.3",
     "@socket.tech/widget": "^0.0.9",
-    "@synthetixio/assets": "^2.0.13",
-    "@synthetixio/contracts": "^1.1.3",
-    "@synthetixio/contracts-interface": "^2.76.1",
-    "@synthetixio/optimism-networks": "^2.74.2",
-    "@synthetixio/providers": "^2.74.7",
-    "@synthetixio/queries": "^3.0.5",
-    "@synthetixio/transaction-notifier": "^2.74.9",
-    "@synthetixio/v3-theme": "^1.0.1",
-    "@synthetixio/wei": "^2.74.2",
+    "@synthetixio/contracts": "^1.1.4",
+    "@synthetixio/contracts-interface": "^2.76.2",
+    "@synthetixio/optimism-networks": "^2.74.3",
+    "@synthetixio/providers": "^2.74.8",
+    "@synthetixio/queries": "^3.0.8",
+    "@synthetixio/transaction-notifier": "^2.74.10",
+    "@synthetixio/v3-theme": "^1.0.2-0",
+    "@synthetixio/wei": "^2.74.3",
     "@tippyjs/react": "^4.1.0",
     "@uniswap/sdk-core": "^3.0.0",
     "@uniswap/v3-core": "^1.0.0",
@@ -141,16 +140,24 @@
     "typescript": "^4.7.4"
   },
   "resolutions": {
-    "synthetix": "^2.0.0",
+    "@synthetixio/contracts": "^1.1.4",
+    "@synthetixio/contracts-interface": "^2.76.2",
+    "@synthetixio/optimism-networks": "^2.74.3",
+    "@synthetixio/providers": "^2.74.8",
+    "@synthetixio/queries": "^3.0.8",
+    "@synthetixio/transaction-notifier": "^2.74.10",
+    "@synthetixio/v3-theme": "^1.0.2-0",
+    "@synthetixio/wei": "^2.74.3",
+    "@types/react": "^17.0.2",
     "@uniswap/sdk-core": "^3.0.0",
     "@uniswap/v3-core": "^1.0.0",
     "@uniswap/v3-periphery": "^1.0.0",
-    "ethers": "^5.6.4",
-    "sharp": "0.29.3",
     "browserslist": "^4.20.4",
     "caniuse-lite": "^1.0.30001355",
+    "ethers": "^5.6.4",
+    "sharp": "0.29.3",
     "shell-quote": ">=1.7.3",
-    "@types/react": "^17.0.2"
+    "synthetix": "^2.0.0"
   },
   "husky": {
     "hooks": {
